requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.25.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.30.0
fake-useragent>=1.4.0